Refactor i2c - move to correct folder - add to Source and Lint .files
