+++ BeagleBone_black/src/Lint.files
@@ -45,6 +45,8 @@ utils/math/quaternion.hpp
 utils/math/vector.hpp
 utils/io/can.hpp
 utils/io/can.cpp
+utils/io/i2c.hpp
+utils/io/i2c.cpp
 utils/logger.hpp
 utils/logger.cpp
 utils/system.hpp
