+++ BeagleBone_black/src/Source.files
@@ -16,7 +16,7 @@ SRCS := \
   utils/concurrent/lock.cpp \
   utils/concurrent/condition_variable.cpp \
   utils/concurrent/barrier.cpp \
-  utils/io/i2c/i2c.cpp \
+  utils/io/i2c.cpp \
   utils/io/can.cpp \
   utils/logger.cpp \
   utils/system.cpp \
